KZ-0 Rename til veilarbfilter
