+++ pom.xml
@@ -4,11 +4,11 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>no.nav.pto</groupId>
-    <artifactId>veilarbfiltrering</artifactId>
+    <artifactId>veilarbfilter</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>no.nav.pto.veilarbfiltrering</name>
+    <name>no.nav.pto.veilarbfilter</name>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
