+++ src/main/kotlin/no/nav/pto/veilarbfilter/service/VeilederFilterService.kt
@@ -1,7 +1,7 @@
-package no.nav.pto.veilarbfiltrering.service
+package no.nav.pto.veilarbfilter.service
 
-import no.nav.pto.veilarbfiltrering.model.EnhetFilterModel
-import no.nav.pto.veilarbfiltrering.model.FilterModel
+import no.nav.pto.veilarbfilter.model.EnhetFilterModel
+import no.nav.pto.veilarbfilter.model.FilterModel
 
 interface VeilederFilterService {
     suspend fun lagreVeilederFilter (veilederId: String, filterValg: FilterModel) : EnhetFilterModel;
